refactor: 엔진바디구동방식 레이블/트림 레이블 컴포넌트 리네이밍, 분리
