+++ front/src/Components/Custom/TrimOptionLabel/DefaultOption.jsx
